requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
betterconf==3.1.0
certifi==2024.7.4
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.8
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
psycopg2==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
typing_extensions==4.12.2
tzlocal==5.2
watchfiles==0.23.0
yarl==1.9.4
pillow==11.0.0
emoji==2.14.0